requirements.txt
python-telegram-bot[ext]
python-dotenv
aiohttp
psutil
aiohttp_cors